7-15-15 finished individual assignment lesson 12

+++ src/citbyui/cit260/thetown/control/ResourcesControl.java
@@ -14,7 +14,7 @@ import the.town.TheTown;
  * @author Eric and Brian
  */
 public class ResourcesControl {
-    public static void addToInventory(int amount, Item type){
+    public static void addToInventory(double amount, Item type){
         Resources[] inventoryList = TheTown.getCurrentGame().getInventory();
         double currentAmount = inventoryList[type.ordinal()].getAmount();
         double total = currentAmount + amount;
